requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncio==3.4.3
bcrypt==4.0.1
cryptography==41.0.5

fastapi==0.104.1
python-dotenv
httpx
numpy==1.26.2
oauth2client==4.1.3
pandas==2.1.3
passlib==1.7.4
pydantic==2.5.1
pydantic_core==2.14.3
PyDrive==1.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium
webdriver_manager
# undetected_chromedriver
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
sqlalchemy 
databases
asyncpg
psycopg2
aiopg
xlrd
xlwt
pdfminer.six
redis
pika